Modified repo for two projects, a parser and a raytracer. Wrote code to parse basic objects
