+++ parser/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "parser"
+version = "0.1.0"
+authors = ["arsenypoga <semiretoja@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "parser"
+path = "src/lib.rs"
+
+[[bin]]
+name = "parser"
+path = "src/bin/main.rs"
+
+[dependencies]
+raytracer = { path = "../raytracer" }
+serde = { version = "*", features = ["derive"] }
+serde_yaml = "*"
+yaml-rust = "*"
+[dev-dependencies]
