+++ raytracer/src/lib.rs
