+++ raytracer/src/render/mod.rs
@@ -1,0 +1,7 @@
+//! This module takes care of all your rendering needs
+pub mod camera;
+pub mod canvas;
+pub mod world;
+pub use camera::Camera;
+pub use canvas::{Canvas, CanvasPart};
+pub use world::World;
