+++ raytracer/src/world/patterns.rs
